some memory optimization work
There's an interesting behavior on startup. On my machine we allocate
around 300 HotplugEvent objects (each of 3604 bytes) on startup and
put them on the queue. When we're done processing an event it's freed.
This leads to interesting memory fragmentation problems. Some
experimental code I've just put in allocates the first 1000 in a fixed
array. Seems to save 200-300kb writable memory.

Also move some bits of HalDevice into HalDevicePrivate. Fix up the
rest of the code not to rely on the ->udi member as it's now only
available through the existing hal_device_get_uid() method.

+++ hald/debug-hald.sh
@@ -3,8 +3,12 @@
 export HALD_RUNNER_PATH=`pwd`/linux:`pwd`/linux/probing:`pwd`/linux/addons:`pwd`/.:`pwd`/../tools:`pwd`/../tools/linux
 export PATH=`pwd`/../hald-runner:$PATH
 
-rm -rf .local-fdi
-make -C ../fdi install DESTDIR=`pwd`/.local-fdi prefix=/
+if [ "$1" = "--skip-fdi-install" ] ; then
+    shift
+else
+    rm -rf .local-fdi
+    make -C ../fdi install DESTDIR=`pwd`/.local-fdi prefix=/
+fi
 export HAL_FDI_SOURCE_PREPROBE=.local-fdi/share/hal/fdi/preprobe
 export HAL_FDI_SOURCE_INFORMATION=.local-fdi/share/hal/fdi/information
 export HAL_FDI_SOURCE_POLICY=.local-fdi/share/hal/fdi/policy
@@ -12,5 +16,5 @@ export HAL_FDI_SOURCE_POLICY=.local-fdi/share/hal/fdi/policy
 echo ========================================
 echo Just type \'run\' to start debugging hald
 echo ========================================
-gdb run --args ./hald --daemon=no --verbose=yes
+gdb run --args ./hald --daemon=no --verbose=yes $@
